fix: load networks on plugin.activate
this fixes the ordering on loading rustyvxcan and docker: docker tries to
connect to rustyvxcan to activate it as a plugin and rustyvxcan tries to
connect to docker to load plugins. This results in retries by docker with
increasing timeouts and thus long boot times.

solution: delay loading of networks in rustyvxcan until docker calls
plugin.activate: rustyvxcan can load and provide endpoints for docker to
connect to and when docker connects rustyvxcan can connect to docker to load
networks.